requirements.txt
aiohttp==1.0.5
async-timeout==2.0.1
certifi==2018.1.18
chardet==3.0.4
ChatterBot==0.8.5
chatterbot-corpus==1.1.2
discord.py==0.16.12
future==0.16.0
idna==2.6
mathparse==0.1.1
multidict==4.1.0
nltk==3.2.5
oauthlib==2.0.7
pymongo==3.6.1
python-dateutil==2.6.1
python-twitter==3.4.1
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
SQLAlchemy==1.2.6
urllib3==1.22
websockets==3.4
wincertstore==0.2
